+++ core/src/main/java/com/nuodb/migrator/jdbc/metadata/inspector/OracleColumnInspector.java
@@ -65,7 +65,7 @@ public class OracleColumnInspector extends SimpleColumnInspector {
         column.setTypeName(typeDescAlias.getTypeName());
 
         int columnSize = columns.getInt("COLUMN_SIZE");
-        column.setSize(columnSize);
+        column.setSize((long)columnSize);
         column.setPrecision(columnSize);
         column.setScale(columns.getInt("DECIMAL_DIGITS"));
